修改字段类型，增加liquibase监听数据初始化
(cherry picked from commit e6d7c69adb7abe2af852b14cb2d51b156c224dbb)

+++ common/risenet-y9boot-common-model/src/main/java/net/risesoft/model/cms/CmsSite.java
@@ -72,12 +72,12 @@ public class CmsSite implements Serializable {
     /**
      * 是否栏目静态页
      */
-    private Byte staticChannel;
+    private Integer staticChannel;
 
     /**
      * 是否文档静态页
      */
-    private Byte staticDoc;
+    private Integer staticDoc;
 
     /**
      * 是否有静态页后缀
